resumes/resumes_app/permissions.py
from rest_framework.permissions import BasePermission

class EditCreateResumePermission(BasePermission):
    def has_permission(self, request, view):
        if request.method == 'POST':
            return request.user.is_authenticated
        return True

    def has_object_permission(self, request, view, obj):
        if request.method in ['PATCH', 'PUT', 'DELETE']:
            return request.user.is_authenticated and request.user == obj.user
        return True