resumes/requirements.txt
asgiref==3.7.2
dj-database-url==2.0.0
Django==4.2.2
djangorestframework==3.14.0
flake8==6.0.0
gunicorn==20.1.0
install==1.3.5
mccabe==0.7.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
